cleanup for consistency
